% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read archival tag sensor data files.}
\usage{
read_file(file, manufacturer, model, sensors)
}
\arguments{
\item{file}{A character with the path to an archival tag data file.}

\item{manufacturer}{A character vector of length 1 with the manufacturer of
the tag.}

\item{model}{A character vector of length 1 with the model of the tag.}

\item{sensors}{A character vector with the sensors embedded in the tags, e.g.
"temperature", "pression". A tag can contain multiple sensors.}
}
\value{
A list with three elements:
\itemize{
\item \code{track_tag_id}: a character with the ID of the tag
\item \code{daylog}: a tibble data.frame with daylog information
\item \code{sensor}: a tibble data.frame with sensor data
}
}
\description{
Read archival tag sensor data files.
}
\examples{
# Read data collected by a CEFAS G5DST tag with pressure and temperature
# sensors
cefas_g5dst_press_temp <- system.file(
  "extdata",
  "cefas_g5dst_pressure_temperature.csv",
  package = "archivaltag"
)
archivaltag::read_file(
  file = cefas_g5dst_press_temp,
  manufacturer = "cefas technology",
  model = "G5dst",
  sensors = c("pressure", "temperature")
)
}
\concept{read functions}
