% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supported_tags}
\alias{supported_tags}
\title{Supported manufacturers, models and sensors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
supported_tags
}
\description{
A look-up table with all manufacturers, models and sensors supported by
archivaltag.
}
\examples{
\dontrun{
# supported_tags.Rda was created with the code below.
supported_tags <- tibble::tibble(
  manufacturer = c(
    "cefas technology",
    "cefas technology",
    "star-oddi"
  ),
  model = c(
    "G5dst",
    "G5pdst",
    "micro-td"
  ),
  sensors = list(
    list("temperature", "pressure", "acceleration"),
    list("temperature", "pressure", "acceleration"),
    list("temperature", "pressure", "acceleration")
  )
)
save(supported_tags, file = "./data/supported_tags.rda")
}
}
\concept{supported tags}
\keyword{datasets}
